+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/mining/multiblockfurnaces/client/MultiBlockBlastFurnaceScreen.java
@@ -7,7 +7,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
 
 public class MultiBlockBlastFurnaceScreen extends AbstractMultiBlockFurnaceScreen<MultiBlockBlastFurnaceMenu>{
-    private static final ResourceLocation TEXTURE = new ResourceLocation(IguanaTweaksExpanded.MOD_ID, "textures/gui/container/blast_furnace.png");
+    public static final ResourceLocation TEXTURE = new ResourceLocation(IguanaTweaksExpanded.MOD_ID, "textures/gui/container/blast_furnace.png");
     public MultiBlockBlastFurnaceScreen(MultiBlockBlastFurnaceMenu pMenu, Inventory pPlayerInventory, Component pTitle) {
         super(pMenu, new MultiBlockBlastFurnaceRecipeBookComponent(), pPlayerInventory, pTitle, TEXTURE);
     }